speech2text-service/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
flask-restx==1.2.0
openai-whisper==20231117
torch==2.0.1
torchaudio==2.0.2
requests==2.31.0
python-multipart==0.0.6
Werkzeug==2.3.7
gunicorn==21.2.0
pytest==7.4.2
pytest-flask==1.2.0
